Move `DsnSyncPieceGetter` to `sync_from_dsn`
